# reference/custom_gpt/participant_action_openapi_v1.yaml
openapi: 3.1.0
info:
  title: AstroHD Participant Interview API
  version: 1.0.0
  description: >-
    Participant-safe API for frozen AstroHD prediction, blind interviewing,
    confirmatory reveal, and explicitly post-hoc exploratory profile ranking.
servers:
  - url: https://YOUR_API_HOST
paths:
  /v1/participant-sessions:
    post:
      operationId: createParticipantSession
      summary: Create and freeze a participant session
      description: >-
        Creates the immutable prediction freeze before behavioral evidence is
        accepted. A scientific-blind GPT should NOT call this operation from the
        participant conversation; use a trusted external intake so the GPT sees
        only the returned opaque session ID. This operation is appropriate for
        self-discovery mode or for the trusted intake application.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BirthIntake'
      responses:
        '200':
          description: Opaque participant session metadata. Hidden chart/predictions are not returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionRecord'
  /v1/participant-sessions/{session_id}/progress:
    get:
      operationId: getParticipantProgress
      summary: Get participant-safe progress
      description: >-
        Safe before reveal. Returns descriptive mapped coverage, unresolved completion
        policy status, and general candidate discrimination, never the true
        birth-state/date rank. Mapped coverage is not a completion criterion.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Participant-safe progress.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProgress'
  /v1/participant-sessions/{session_id}/next-question:
    get:
      operationId: getParticipantNextQuestion
      summary: Select the next neutral interview question
      description: >-
        Uses answer-blind candidate discrimination to suggest an informative
        dimension. The interviewer may clarify a mixed answer before moving on.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Suggested question or open-narrative guidance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NextInterviewQuestion'
  /v1/participant-sessions/{session_id}/evidence:
    post:
      operationId: appendParticipantEvidence
      summary: Append one atomic participant observation
      description: >-
        The backend assigns the current phase. Only trait/behavior observations
        with a valid frozen question token can affect natal ranking. Outcome,
        timing, environment and conventional covariate evidence is retained for
        separate research layers. Use answer=null when an Other/free-form answer
        cannot honestly be mapped to a frozen option.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceInput'
      responses:
        '200':
          description: Immutable appended evidence record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicEvidenceRecord'
  /v1/participant-sessions/{session_id}/lock:
    post:
      operationId: lockParticipantConfirmatoryEvidence
      summary: Lock pre-reveal confirmatory evidence
      description: >-
        Fails with SCIENTIFIC_COMPLETENESS_POLICY_UNRESOLVED until an owner-explicit,
        server-owned completion policy is authorized. Mapped coverage cannot authorize
        a lock.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Confirmatory evidence-lock metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicConfirmatoryLock'
  /v1/participant-sessions/{session_id}/reveal:
    post:
      operationId: revealParticipantResult
      summary: Reveal frozen predictions and confirmatory rank
      description: >-
        Available only after a conforming lock bound to an authorized server-owned
        completion policy. A new reveal is unavailable while owner authority is
        unresolved.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Confirmatory reveal report.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /v1/participant-sessions/{session_id}/finalize-exploratory:
    post:
      operationId: finalizeParticipantExploratoryProfile
      summary: Finalize the post-hoc holistic profile and ranking
      description: >-
        Computes a second ranking from the final refined profile after reveal.
        It is explicitly marked post-hoc/exploratory and does not overwrite or
        count as independent confirmation of the pre-reveal result.
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Final report containing confirmatory and exploratory rankings side by side.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /v1/participant-sessions/{session_id}/final-report:
    get:
      operationId: getParticipantFinalReport
      summary: Fetch the completed participant report
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Completed confirmatory plus exploratory report.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^HD-[A-F0-9]{32}$'
  schemas:
    BirthIntake:
      type: object
      additionalProperties: false
      required:
        - local_datetime
        - birthplace
        - iana_timezone
      properties:
        local_datetime:
          type: string
          format: date-time
          description: >-
            Local civil wall-clock datetime with no UTC offset. Example:
            1994-01-28T00:35:00.
        birthplace:
          type: string
        iana_timezone:
          type: string
          description: Historical IANA zone such as Europe/Istanbul.
        fold:
          type:
            - integer
            - 'null'
          enum: [0, 1, null]
          default: null
          description: Required only when the local civil time is ambiguous.
        mode:
          type: string
          enum: [scientific_blind, self_discovery]
          default: scientific_blind
        ranking_scope:
          type: string
          enum: [known_birth_month, century_global]
          default: known_birth_month
          description: >-
            known_birth_month is implemented in v1. century_global is reserved
            and fails closed until the reusable arbitrary-participant century
            candidate cache is certified.
    SessionRecord:
      type: object
      additionalProperties: true
      required: [session_id, mode, ranking_scope, prediction_freeze_sha256]
      properties:
        session_id:
          type: string
          pattern: '^HD-[A-F0-9]{32}$'
        mode:
          type: string
          enum: [scientific_blind, self_discovery]
        ranking_scope:
          type: string
          enum: [known_birth_month, century_global]
        prediction_freeze_sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
    EvidenceInput:
      type: object
      additionalProperties: false
      required: [domain, narrative]
      properties:
        domain:
          type: string
          enum:
            - trait
            - behavior
            - outcome
            - timing
            - environment
            - conventional_covariate
        question_id:
          type: [string, 'null']
          default: null
        answer:
          type: [string, 'null']
          default: null
          description: >-
            A frozen response token only when it genuinely fits. Null means the
            narrative is retained without forcing a scoring category; for a
            previously scored question it neutralizes that earlier token.
        behavioral_confidence:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5
        measurement_reliability:
          type: number
          minimum: 0
          maximum: 1
          default: 0.75
        narrative:
          type: string
          minLength: 1
        childhood_pattern:
          type: [string, 'null']
          default: null
        adult_pattern:
          type: [string, 'null']
          default: null
        contexts:
          type: array
          items: {type: string}
          default: []
        exceptions:
          type: array
          items: {type: string}
          default: []
        example_text:
          type: [string, 'null']
          default: null
        counterexample_text:
          type: [string, 'null']
          default: null
        minimum_evidence_passed:
          type: boolean
          default: false
          description: >-
            True only when the selected question's minimum-evidence requirement
            has actually been met.
        consistency_status:
          type: string
          enum: [not_checked, consistent, reconciled, unresolved]
          default: not_checked
        quality_rationale:
          type: [string, 'null']
          default: null
          description: >-
            Evidence-specific reason for passing the minimum and consistency gates.
            Required by the server for adequate trait/behavior evidence.
        supersedes_evidence_id:
          type: [string, 'null']
          default: null
    PublicEvidenceRecord:
      type: object
      additionalProperties: false
      required: [schema_version, evidence_id, session_id, phase, created_at_utc, evidence]
      properties:
        schema_version: {type: string, const: participant-public-evidence-v1}
        evidence_id: {type: string}
        session_id: {type: string}
        phase:
          type: string
          enum: [confirmatory_blind, posthoc_exploratory]
        created_at_utc: {type: string, format: date-time}
        evidence:
          $ref: '#/components/schemas/EvidenceInput'
    PublicConfirmatoryLock:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - session_id
        - locked_at_utc
        - evidence_count
        - scoring_response_count
        - scoring_responses_sha256
        - excluded_non_natal_evidence_count
      properties:
        schema_version: {type: string, const: participant-public-confirmatory-lock-v1}
        session_id: {type: string}
        locked_at_utc: {type: string, format: date-time}
        evidence_count: {type: integer, minimum: 0}
        scoring_response_count: {type: integer, minimum: 0}
        scoring_responses_sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        excluded_non_natal_evidence_count: {type: integer, minimum: 0}
        evidence_quality_contract_version: {type: [string, 'null']}
        adequately_assessed_question_count: {type: [integer, 'null'], minimum: 0}
        required_question_count: {type: [integer, 'null'], minimum: 0}
        completion_policy_id: {type: [string, 'null']}
        completion_policy_digest:
          type: [string, 'null']
          pattern: '^[a-f0-9]{64}$'
        completion_authority_source_ref: {type: [string, 'null']}
    PublicProgress:
      type: object
      additionalProperties: false
      required:
        - session_id
        - phase
        - confirmatory_observation_count
        - scoreable_observation_count
        - non_natal_observation_count
        - mapped_scoreable_question_count
        - mapped_scoreable_coverage
        - adequately_assessed_mapped_question_count
        - completion_policy_status
        - completion_policy_id
        - completion_required_question_count
        - completion_coverage
        - completion_authority_source_ref
        - true_birth_rank_concealed
      properties:
        session_id: {type: string}
        phase:
          type: string
          enum:
            - confirmatory_blind
            - confirmatory_locked
            - revealed
            - posthoc_exploratory
            - finalized
        confirmatory_observation_count: {type: integer, minimum: 0}
        scoreable_observation_count: {type: integer, minimum: 0}
        non_natal_observation_count: {type: integer, minimum: 0}
        mapped_scoreable_question_count:
          type: integer
          minimum: 0
          description: Descriptive count from the frozen mapping artifact, not policy.
        mapped_scoreable_coverage:
          type: number
          minimum: 0
          maximum: 1
          description: Descriptive mapped-artifact coverage, not scientific completeness.
        adequately_assessed_mapped_question_count: {type: integer, minimum: 0}
        completion_policy_status:
          type: string
          enum: [UNRESOLVED_OWNER_AUTHORITY, AUTHORIZED, REVOKED, STALE]
        completion_policy_id: {type: [string, 'null']}
        completion_required_question_count: {type: [integer, 'null'], minimum: 0}
        completion_coverage: {type: [number, 'null'], minimum: 0, maximum: 1}
        completion_authority_source_ref: {type: [string, 'null']}
        candidate_state_count:
          type: [integer, 'null']
        top_state_tie_count:
          type: [integer, 'null']
        top_margin_rubric_bits:
          type: [number, 'null']
        true_birth_rank_concealed:
          type: boolean
          const: true
    NextInterviewQuestion:
      type: object
      additionalProperties: true
      required: [session_id, prompt, response_format, allow_other, guidance]
      properties:
        session_id: {type: string}
        question_id:
          type: [string, 'null']
        prompt: {type: string}
        response_format: {type: string}
        followups:
          type: array
          items: {type: string}
        expected_information_gain:
          type: [number, 'null']
        adjusted_utility:
          type: [number, 'null']
        allow_other:
          type: boolean
          const: true
        guidance: {type: string}